Add OXT Token (#791)
* Add OXT Token

This change adds the Orchid OXT ERC-20 token.

* Remove Incorrectly Named Directory to Fix Tests

This directory is not named using the checksum address. There is a different directory for this token that is correctly named. However, the presence of this old directory is causing the tests to fail. Removing it should hopefully resolve this issue.

* Remove another incorrectly named (duplicate) dir

There is another duplicate incorrectly named directory that is not in checksum format and is causing the tests to fail.

* Remove a third incorrectly named duplicate dir