Trips token
Trips, is TripsCommunity's cryptocurrency, an ERC-20 token. It is distributed regularly or as a one off 'payment' as a reward for completed tasks to community members who contribute to the project. It is also used internally between the community members to pay for goods or services (i.e. a stay in one of TripsCommunity vacation rental properties).