NS7/info
Najah Safiya (NS7) is a standard ERC20 token. The primary goal of this Project is to function as a coin used in a specific android/iphone game (currently un-named however thought to be eponymous in part, projected to be available on Google Play and Apple iTunes apps. The coin (NS7) will also function as a standard coin tradable on various exchanges for other cryptocurrencies like Bitcoin, Ethereum, 0x, Stellar Lumens, Salt, BAT, etc. Najah Safiya's end goal is to function as a standard coin and enlarge the cryptospace by reaching an active group of consumers -- gamers -- who may not currently be as aware of crypto or into trading as yet.