Limited Financial Logo And Website information upload Trust wallet
Limited.Finance LIFI is a community managed token. keep tokens with limited amount. Totally transparent, making sure that price of the token will not be dropped in a future and become one of the best deflationary crypto in the inflation financial world. Total supply 14,300.  Reserved 1000 - future dev. marketing and bounty. 4,300 - Rewarding, DAO and staking.