Files about Green Energy Coin
Files about Green Energy Coin.
The Green Coin is a by-product of Euro Energy S / A and was created as a new asset to support an electromagnetic solar generator project in the sustainable energy market. Green Energy Coin claims to be backed by Green Bonds Issuance and Reduced Emission Certificates.