Add BnEX Token (BNX)
The Bnex project is a decentralized financial platform based on a Binance smart chain. Collaborate with BSC to advance Bnex projects. The goal of governance mechanisms is to gradually shift from Proof-of-Authority (POA) to DAO and ultimately run by the community. The Bnex project uses the Binance Defi ecosystem program to build a decentralized community.

Bnex does not have pre-mining and team-held tokens and distributes all BNX based on contributions. Initially, the Bnex team decides on distribution and proposes and accepts governance over the long term through DAO. For example, you can vote on key topics such as token economy, issuance limits, improvements and changes, and receive BNX for participation in governance.