Added Dondi Logo and JSON info file
We have about 1000 users on our platform (feel free to verify on https://dondi.io/), but they have not all performed the "Harvest" function on our drip system. We will certainly pass 500 holders on Etherscan when our liquidity pools open on the 30th and DONDI is required for our LP. We are hoping to have our logo on Uniswap up before then. 

Thank you!