STACY Token Assets
Hi, 

Submitting on behalf of Chads.vc team. This is the Chadswap governance token. 

We are aware of the 500 token holder requirement, we expect to reach that tomorrow when Uniswap trading opens. (A bunch of fake clones have launched and we need a picture to identify the real STACY)

Our CHADS token has about 2000 holders.

Thank you