Add XBT tokens information
Hello,

I'm proposing the new tokens to the crypto space. It is called Elastic Bitcoin or XBT.

XBT is a DeFi protocol with an elastic supply model. This means the total supply of XBT is constantly changing and achieved by utilizing the power of Ethereum's Smart Contract to adjust supply based on market condition.