Add XBT token information
Hello,

I'm proposing a new Defi protocol to the crypto space, which is called Elastic Bitcoin or XBT.

XBT is a DeFi protocol with an elastic supply model. This means the total supply of XBT is constantly changing and achieved by utilizing the power of Ethereum's Smart Contract to adjust supply based on market condition.