Update XSGD logo
Hello there,

Nice to meet you. I am JJ, handling marketing for Xfers. Since our previous message we released the XSGD token to the public on October 5th:

https://cointelegraph.com/news/crypto-singapore-dollar-aims-to-diversify-landscape-dominated-by-usd
https://medium.com/xfers-sg/introducing-xsgd-the-singapore-dollar-backed-and-travel-rule%C2%B9-compliant-stablecoin-753ea48c7b8e
https://finance.yahoo.com/news/xfers-announces-public-release-singapore-060000153.html

As a reminder, XSGD is the first regulated Singapore dollar-backed stablecoin, it runs both on the Ethereum blockchain as an ERC-20 token and on the Zilliqa blockchain as a ZRC-2 token.

Since launch, XSGD has been transacted >10,000 times for >3,400 holders on the Zilliqa blockchain and >100 times for ~500 holders on the Ethereum blockchain.

https://viewblock.io/zilliqa/address/zil180v66mlw007ltdv8tq5t240y7upwgf7djklmwh?txsType=token-contract&specific=zil180v66mlw007ltdv8tq5t240y7upwgf7djklmwh
https://etherscan.io/token/0x70e8de73ce538da2beed35d14187f6959a8eca96

We believe that one of the reasons why XSGD is not growing as fast on Ethereum as on Zilliqa is that we haven't been able to update the logo to its new logo on Trustwallet, notably because this logo database is also used by Uniswap. Hence, users can't recognise XSGD on Uniswap and don't trade it as the XSGD pairs seem fraudulent.

We are hoping that you will be understanding of this situation and enable us to update the XSGD logo in your DB.

Best Regards,
Liew Jia Jun