Addition of OML to Trust Wallet (#5279)
* Addition of OML to Trust Wallet

I am aware of the current review requirements, however, please consider reading the details below;

Initially when we started our project, requirement for addition to Trust Wallet was 1000 addresses + 2500 transactions (no mention of airdrop tokens). We knew it was doable, so we started recommending Trust Wallet to our customers and airdrop participants. We have even prepared a guide that explains how to use Trust Wallet and how to add our token manually. Here is the link: https://www.youtube.com/watch?v=PL6llNer22s

The video has over 11K views, and more than 7000 people downloaded Trust Wallet just because we recommended and they are airdrop participants. We had a list of 25000 users for airdrop, however, in order to verify that every registration is valid and users are real, we asked for phone verification and we ended up with around 9000 verified airdrop participants. 

The full list of addresses could be reached at https://omlira.com/airdroplistesi.txt - We grouped 50 addresses together, announced it on our socials (example: https://twitter.com/om_lira/status/1348172409342451712?s=20). Our social accounts, followers, and telegram channel (https://t.me/omlira) could be seen.

We have airdropped to more than 5000 participants now and we will airdrop to around 4000 more accounts. We have already paid more than 6 ETH for airdrop so far. So, we believe that OML could be added to the TrustWallet simply because we are not trying to scam anyone or cause problems but just trying to make Trust Wallet usage easier for our customers. 

We could make any payment necessary to add OML to Trust Wallet.

* Delete blockchains/ethereum/assets/0x224db5e6180761df4c3d8936585f6b8b83879770 directory

* Correction of the directory name