Add USxD token
USxD is an acronym that stands for United Stock exchange Dollar. 
United Stock was chosen to represent the unification of different sectors, that of cryptocurrencies and stock indices, and the different platforms on which these different markets operate. United Stock emphasizes the very purpose of USxD which is to become a medium of exchange created to trade Crypto Stocks but which aims to become universal.USxD is an acronym that stands for United Stock exchange Dollar. 
United Stock was chosen to represent the unification of different sectors, that of cryptocurrencies and stock indices, and the different platforms on which these different markets operate. United Stock emphasizes the very purpose of USxD which is to become a medium of exchange created to trade Crypto Stocks but which aims to become universal.