Added STAKD Finance Token Logo
Added STAKD Finance Token Logo.

We are a powerful defi ecosystem on binance smart chain that wants to revolutionize the growing bsc sector by providing all in one solution for everyone. We are aiming to create a solid competitor to ethereum financial products and want to be an integral part of binance smart chain ecosystem.