add new logo for megaBONK
Bonk Token, which is already in the assets section for Trust Wallet is migrating to a new contract address in this file. 

https://bonktoken.medium.com/migration-to-megabonk-bc1ffc5dd9d0

Here is a link from the official Bonk token medium that describes the migration upgrade 

Knowing this, I hope we can waive the listing fee & accept the new logo even though qualifications for Trust Wallet have changed.