add BitBlocks Finance v2 (BBKFI)
BitBlocks Finance v2 (BBKFI) is a token of the BitBlocks Project owner of https://bitblocks.finance

In addition to it, we have our own cryptocurrency from 2018 called BBK (Quark Algorithm) and the Wrapped BBK (BEP-20)

Our project website is https://bitblocksproject.com/en-us and our coin explorer distribution shows that we have more than 19000 BBK holders https://bbk.ccore.online/distribution

We also have in our ecosystem, an e-sports team and a cryptocurrency payment application.

Although BBKFI is a relatively new currency, we are already listed on Blockfolio, DappRadar, Audit on TechRate and others coming soon, all the information filled in at BSC Scan and over 18,000 transactions, despite still having a number of holders that do not meet the 2500 requested, our TVL is concentrated in Farms and Pools, as well as holders.

For this reason, we ask you to release the minimum holders requirement so that our token can be listed on the Trust Wallet. If our project can be listed, we request the ok to proceed with the payment.