Add Official Logo for PooCoin (#6539)
* Add PooCoin Logo and Info

This is the logo and information for PooCoin, a reflect token on the Binance Smart Chain. The token is related to Poocoin.app, a tracking and charting website for BSC wallets and assets.

* Add files via upload

* Delete PooCoin_Logo.png

* Add PooCoin Logo and Info

PooCoin logo and info. PooCoin is a reflect token on the Binance Smart Chain. It is directly related to Poocoin.app, a charting and wallet tracking website for the Binance Smart Chain.

* Add PooCoin Info and Logo

PooCoin logo and information. PooCoin is a reflect token on the Binance Smart Chain.  PooCoin is directly related to Poocoin.app, a charting and wallet tracking website for the Binance Smart Chain.

* Add PooCoin Info and Logo

PooCoin logo and information. PooCoin is a reflect token on the Binance Smart Chain.  PooCoin is directly related to Poocoin.app, a charting and wallet tracking website for the Binance Smart Chain.

* Delete info.json

* Add PooCoin Info and Logo

PooCoin logo and information. PooCoin is a reflect token on the Binance Smart Chain.  PooCoin is directly related to Poocoin.app, a charting and wallet tracking website for the Binance Smart Chain.

* Add PooCoin Info and Logo

PooCoin logo and information. PooCoin is a reflect token on the Binance Smart Chain.  PooCoin is directly related to Poocoin.app, a charting and wallet tracking website for the Binance Smart Chain.

* Add PooCoin Logo and Info

This is the logo and information for PooCoin, a reflect token on the Binance Smart Chain. The token is related to Poocoin.app, a tracking and charting website for BSC wallets and assets.

* Delete blockchains/smartchain/assets/0xB27ADAfFB9fEa1801459a1a81B17218288c097cc directory

* Delete blockchains/smartchain/assets/0xb27adaffb9fea1801459a1a81b17218288c097cc directory

* Add PooCoin Logo and Info

This is the logo and information for PooCoin, a reflect token on the Binance Smart Chain. The token is related to Poocoin.app, a tracking and charting website for BSC wallets and assets.

* Add PooCoin Logo and Info

This is the logo and information for PooCoin, a reflect token on the Binance Smart Chain. The token is related to Poocoin.app, a tracking and charting website for BSC wallets and assets.

* Add PooCoin Logo and Info

This is the logo and information for PooCoin, a reflect token on the Binance Smart Chain. The token is related to Poocoin.app, a tracking and charting website for BSC wallets and assets.

* Add Official Logo for PooCoin

Replacing the stock image originally used with an official, community created icon.

* Add Official Logo for PooCoin

Replacing the stock image originally used with an official, community created icon.

Co-authored-by: Catenocrypt <catenocrypt@users.noreply.github.com>