Added Logo and token information for moonclone
I am submitting logo and information for moonclone. MOONCLONE is the first fully autonomous DeFi project building ecosystem. Our first release is a token creation bot called MOONCLONE BOT. Teams can purchase a SAFEMOON clone contract, or a custom tokenomics contract by answering a few questions in the MOONCLONE BOT on Telegram. Tokens are deployed to the wallet that purchased the contract. Instructions are then sent on how to add & lock liquidity, and renounce ownership. TOKENPAY is a telegram payment tool that accepts any cryptocurrency listed on pancakeswap to be accepted for products and services. When purchasing MOONCLONE products & services, tokens are traded into MOONCLONE where 5% is added back to liquidity, and 5% is redistributed to holders. TOKENPAY will be available for integration with any telegram channel!