RoyalDoge is a frictionless yield generation, hard-coded charity-based token on the Binance Smart Chain
Losing even one single species can have disastrous impacts on the rest of the ecosystem, because the effects will be felt throughout the food chain. From providing cures to deadly diseases to maintaining natural ecosystems and improving overall quality of life, the benefits of preserving threatened and endangered species are invaluable.
So, we have decided to create a great performing token with a limited supply, offering 5% transaction fees, where 2.5% will get burned and 2.5% will be donated to preserve wildlife and contribute in helping save our planet.