Add Moonelon logo and info
In this opportunity we're adding the logo and basic information of Moonelon.
 
What is Moonelon?
Think of Moonelon as a fair universal currency for the people. Everywhere. It is not just another meme currency like Doge, Shiba, or dogelon. Sure they are fun but why not have fun while building something awesome at the same time?, This is a real project that will be constantly evolving. And it begins with a great community full of hope that can only dream of making it to the moon. You may have some questions.