Logo and Info.json SafePizza
Added info and logo for SafePizza to merge and show it on trust wallet. 

We have over 7k holder and more than 29k transfers. 

The community asks every day about the trust wallet logo and now we want to get them what they ask us for, so please bring it online =)