Added Official Logo of ElonTech (ETCH) Token
I have added the logo of ElonTech to TrustWallet as described by the community guidelines of the logo. Here is the Token contract address

https://bscscan.com/token/0xc66c8b40e9712708d0b4f27c9775dc934b65f0d9
Here is the website: https://elontech.finance

Kindly approve it. All requirements are fulfilled. Our community is looking for a logo in their wallets.

Thank you
Jack