Add MUTT Logo and Info
We are a group of U.S. Marine Corps veterans who created a token to help raise money for various veteran organizations. We are running a legit token and want to have nothing but the utmost integrity with our project and we believe adding an image to our token will add a level of trust between us and potential new investors. Thank you.