Add Pendle token logo and info
The token requested to be featured here is Pendle's native governance token. Pendle is a protocol that enables the trading and hedging of yield. The trading occurs on Pendle AMM that is designed for assets that are time-dependent. Pendle token is also part of Sushiswap's Onsen program with good liquidity and exposure.