Add Max Crowdfund (MCF) logo and info
We have changed from our ardor child chain (Max Property Group - MPG) to the Binance Smart Chain and created the Max Crowdfund - MCF token. This is a new token, but we have a large community and approx 60,000 token holders.
We are going through the steps (process) to create the token, add the token to Trust Wallet, add Liquidity to Pancakeswap and informing our community (https://twitter.com/maxcrowdfund) about the changes. We will slowly be distributing the MCF tokens to all our users and trading will commence.

We understand you do not like new tokens, but this is just us changing. Our group was incorporated in 2016 in the Netherlands and a lot of information can be found online about Max Property Group and Max Crowdfund.