Add RST logo and info
Add logo and information for Realio Security Token. RST is issued to represent ownership of the Realio Ecosystem. The Realio Security Token (RST) represents an investment in the platform and ecosystem and offers a revenue stream in the form of profit share. RST was sold through a security token offering (STO).