Logo Update (#10011)
* Add Travel Care Logo in Trust Wallet

* Upload TWT Logo

Hello,
In the trust wallet the logo is cut off, I have sent the logo with the exact size indicated in the instructions (256x256)
But since there is no margin, the logo is cut off.  i already pay fee yesterday and the PR was #9976
Here i attached a logo with margins. Many Thanks

Co-authored-by: Andrew M <35627271+zachzwei@users.noreply.github.com>