Add BITE logo and info
Project background:

Official announcement:
https://t.me/DragonBiteD8

Project Website:
https://dragonbite.io/

Project team background:
DragonBite is founded by a talented team of blockchain enthusiasts. The team has accumulated over 50 years of experience in digital marketing, technology, loyalty, payments, banking, consulting and fintech. The DragonBite team is based in China, Hong Kong, Singapore, the United Kingdom and the United States.

Project basic information:
DragonBite is the solution to a massive and global problem: there is more value in loyalty points, rewards and coupons in circulation in the world than US dollars, yet the majority of these digital assets are idle. The lack of a unified platform for users to redeem and spend their loyalty points means that every year hundreds of billions of dollar worth of assets become stale.

This is where we come in DragonBite is a universal digital asset management platform powered by Blockchain technology. We leverage cryptocurrencies and DeFi to help loyalty points holders consolidate their earnings, interact with a global ecosystem of merchants and unlock the value of their digital assets.

As the vision of DragonBite is to disrupt the loyalty industry using blockchain and DeFi. We aim to remove the intrinsic inefficiencies of loyal points by allowing loyalty points turn into cryptocurrency. The whole idea is to facilitate loyalty points’ owner.

95% of the population owns loyalty point but the adaption rate of cryptocurrency is only 5%. DragonBite provides an easy and straight forward way for non-crypto users embrace the crypto space.

Media publications:
https://cointelegraph.com/press-releases/dragonbite-raises-over-27-million-within-48-hours
https://finance.yahoo.com/news/cryptocurrencies-token-bite-announces-first-150600032.html
https://www.jinse.com/news/blockchain/1097679.html

Tradeable exchanges:
UNISWAP (From 11/06/2021)

Recommended Gas Limit (For ERC-20 project teams) for transaction: 60,000