Add token request - Wrapped ILCOIN - WILC
Hi trustwallet.com community we would like to be added to your repo in order to have our info and logo updated on Uniswap and others, our token WILC is 1:1 with our crypto project ILCOIN that is a SHA-256 proof of work and now we are providing our community with the possibility of tokenization of our crypto by giving them the possibility of participate in Defy projects. Our Wrapped ILCOIN (WILC) token works similar to WBTC and WETH.

Thank you for your great efforts and best regards!.
ILCOIN Development Team.