Tennisball Token TBT
The project maintains the long-term goal of launching a Tennis App, providing a space for Tennis shop, sales and sharing, as well as other utilities and sub-projects that are yet to be announced. The strongest asset of Tennisball is the incredible community that has been built around it. Giveaways & competitions are an integral part of Tennisball.