Add PING  (BEP20) Logo and Info
PING Token is from the Sonar Platform on Binance Smart Chain(BSC)
Sonar is building the next-gen crypto tracking dashboard. It is the most complete and easy to use tracking and analytics tool for the BSC and ETH networks.
https://www.sonarplatform.io/

This Pull request is to update PING token logo on Trustwallet.