Shrimps Token add asset
SHH incentivises humans to engage in environmentally friendly and ecologically sustainable acts in order to promote marine conservation and offset carbon emissions with algae farms.  SHH can be spent in participating merchants and offers a redemption platform app for holders of SHH to use the token in meaningful ways.  The goal is to provide a green economy activation engine that allows people to earn Shrimps through the earn&build program and provide a form of UBI for everyone to participate.  We updated our token to a new address due to a security issue which has since been corrected. 

Our team is made up of conservation and ESG experts who have been engaged in SDGs for over 20 years.  We have a balanced team of legal, security, conservation, and marketing advisors that drives the adoption of the token forward.  Please see our lite paper for more information.