Add files via upload
We have been working hard with our community to meet the minimum requirements to have our logo and price show up in the Trust Wallet. This is a big milestone for us in our project. We have been working very hard to grow our community and would love to complete this milestone.  Thank you for the consideration.