Deed Token
The DEED Token is described to be a crowdfunding ecosystem that works for financial inclusion and social equality, generating opportunities through social programs, encouraging sustainable, long-term practices to promote community growth and development.

DEED is a secured and efficient tradeable cryptocurrency and donation coin committed to building a unique and secure system aimed at assisting and financially supporting charitable organizations that seek to engage the next generation of philanthropists through the use of blockchain technology. DEED designed not only to take charity, to the next level but also to help individuals who took part in our ICO increase and multiply their investment.

DEED is expected to be the next major cryptocurrency. It is currently an ERC20 token, but will subsequently be upgraded to it’s own chain.