Please add logo to Trust
Please add logo to Trust. 

@cryptocool @cryptomanz

Please apply payment with memo 1110 to this request.

We are pretty sure this file will work now. We have already paid with memo 1110 that has never been applied to a PR:

https://explorer.binance.org/tx/5B3AEE89F6697B096A7FE232D945D0A7218A8FCABB8E873511CCF318FCCAFE5C

Please find this payment and apply to this request so we can get this accomplished successfully.