Add Kirobo DAPP Image
Kirobo is a leading developer of innovative blockchain solutions tailor-made for the DeFi ecosystem.
Kirobo's mission is to make cryptocurrency an everyday tool that everyone can use with confidence. 
We achieve this with user-friendly products that remove the risk of human error and fraud from cryptocurrency transactions, making crypto management as safe as can be. Kirobo is backed by the proud recipient of two grants from the Israeli Innovation Authority — the only blockchain technology company to achieve this.