Add TruBadger (BEP20) (#11894)
* Add logo to Trust Wallet

Please add our token logo to Trust Wallet. We have achieved all requirements.

* Add files via upload

* Add files via upload

* Add Logo to Trust

Please add our Logo to Trust. I believe we have everything required but please do let me know if not.

* Please add our Logo to Trust

This is the 2nd request today. Previous was closed with no explanation so I do not understand. Please help us.

* Please add logo to Trust

We have sent 1000 TWT-BEP2 to the proper address with memo 1110

* Please upload our logo to Trust - Payment Sent

We have already paid 1000 TWT with memo we had at the time of 1110. We have had multiple submissions trying to fix this. We think we finally have it right. Please apply our previous payment to this request as we have not had a successful completion of a request as of yet. Same info, same token, same logo.

Payment of 1000 TWT-BEP2 has been sent with the memo 1110 https://explorer.binance.org/tx/5B3AEE89F6697B096A7FE232D945D0A7218A8FCABB8E873511CCF318FCCAFE5C

* Please add logo to Trust

Please add logo to Trust. 

@cryptocool @cryptomanz

Please apply payment with memo 1110 to this request.

We are pretty sure this file will work now. We have already paid with memo 1110 that has never been applied to a PR:

https://explorer.binance.org/tx/5B3AEE89F6697B096A7FE232D945D0A7218A8FCABB8E873511CCF318FCCAFE5C

Please find this payment and apply to this request so we can get this accomplished successfully.

* Add Logo to Trust - CMC Listed

@cryptomanz

Payment is done and status was on hold for cmc listing: 

Check PR #11543

TruBadger is listed:

https://coinmarketcap.com/en/currencies/trubadger/

https://www.coingecko.com/nl/coins/trubadger

Can you approve and procces.

Thanks.

* Logo updated transp background

* logo white background deleted

* Info for token TruBadger

* Delete logo.png

* Update info.json

* Delete info.json

* Delete info.json

* Delete logo.png

Co-authored-by: TruBadger <trubadger@users.noreply.github.com>
Co-authored-by: CryptoA <kulishnastia200@gmail.com>