Adding PEEPS to TrustWallet (#12009)
* Adding PEEPS to TrustWallet

PEEPS is a community-driven coin built by people within that very same community. It will host an ever-adapting ecosystem built with the visibility and security of the ETH network. The transparency and welcoming community will both allow organic growth and the promotion of ideas to improve the existing crypto landscape.

* Adding PEEPS to TrustWallet

PEEPS is a community-driven coin built by people within that very same community. It will host an ever-adapting ecosystem built with the visibility and security of the ETH network. The transparency and welcoming community will both allow organic growth and the promotion of ideas to improve the existing crypto landscape.
Commit directly to the master branch.

* Update info.json

Co-authored-by: Nazar <85495047+cryptomanz@users.noreply.github.com>