I am adding the logo and info for DrunkDoge token!
This request is about the initial logo and information addition of our token DrunkDoge. Our token is popular, well known and has great mechanics. We are on CMC and CG. The reason that I submit this logo and info is to prevent people from getting scammed and provide initial information about our token.