DogeMoon is one of the first charity tokens on the BinanceSmartChain.
It had both main features of typical RFI forks, which are auto-burns to locked liquidity (2%) and auto-farming mechanics (2%).
Dogemoon will be using the predominant amount of the transaction tax to fund the Marketing/Charity-wallet. The 6% transaction towards this wallet is auto-turned into BNB to prevent a major sell-off if we use significant amounts. We ALWAYS discuss the milestone donations with the community.
https://solidity.finance/audits/DGMOON/