Our gaming ecosystem
GAMEX is a project with goal of creating a gaming ecosystem revolving around the benefits of the players. Here at GAMEX we are determined to succeed in uniting all gamers around a platform that can help them succeed and reach higher levels in their careers. With the popularity of in-game-NFTs. the worlds of gaming and blockchain have never been so connected, and we are providing gamers with a friendly ecosystem, built to sustain AAA games and indie games alike.