Sombra - SMBR logo and json info upload re-attempt
Uploading the Sombra Network logo and Json. We have fixed the non checksummed and duplicate ID errors in the json file. Thank you so much. 

Please let us know if the payment is something we need to do as we were unfamiliar with that. Our token previously was Cheecoin and we have now moved contracts and re-branded with the same community and team. Cheecoin was already on Trust. We are fine to pay if needed but wanted to check that we met the requirements as SMBR prior to sending the non refundable amount. 

Cheers,
Brendan and Sombra