Add files via upload
Changes made to naming and info.
Although we are not at the required holders threshold, however, over the past few days, we have done quite a lot to proof the legitimacy of the project, among which is Team being Kyced, listing on multiple exchanges and analytic platforms such as CG, nomic etc. We would appreciate if our request is considered and granted.

Thanks.