Adding info and logo for SpacePortCoin (SPC)  (#13338)
* Adding info and logo for SpacePortCoin (SPC)

Currently SPC is listed in CMC, but in TrustWallet logo is not shown. Hence following the suggested steps to update the logo for TW for SPC.

* deleting to provide correct format

* Adding info and logo for SpacePortCoin (SPC)

Currently SPC is listed in CMC, but in TrustWallet logo is not shown. Hence following the suggested steps to update the logo for TW for SPC.

* Updating social media links

Updating social media links as suggested

* Updating links as suggested

Updating format of links and removing www as suggested

* Updating discord link for better visibility

Although the same server, updating the discord link with official name