Adding Logo and Information for the Token
Romeo Doge is a community driven reward token with built in feature to reward all holders with multiple cryptocurrencies. 
Romeo Doge holders will receive multiple rewards on the BEP20 network (Binance Smart Chain). Our smart technology studies the market trend and rewards it's holders with a different cryptocurrency that has potential of fast growth on every reflection. 
Romeo Doge is designed to build serious wealth for investors. It does this by applying a 18% tax on transactions. Out of this, 10% gets redistributed to holders in binance pegged coins, 3% goes to the liquidity pool to build a stronger price floor for all investors and 5% goes to the marketing wallet to fuel consistent growth.