Add PALDOG logo and info
PalmaDOGE  is the new project which paves the way for a series of [PALMA] projects we will carry out in the next 3 years. 
With unlimited marketing budget & clearly roadmap, we will synchronize the features of traditional and new scheme. 
With PALDOG, you will experience a new chain of features from Stake/Farm,NFT Game/Marketplace, Jackpot, Swapmarket ,
 App, Launchpad, and finally Blockchain as the way DOGECoin did