[Urgent] Please change LOGO and REMOVE INFO OLD X2P. (PAID) #13593
Please CHANGE the old X2P logo with the logo that i have included in the folder. I have already paid for this to be done together with the upload with new X2P logo. Please change the old logo as people are still buying through old contract. Thanks in advance!