[Urgent] Please change LOGO and REMOVE INFO OLD X2P. (PAID) #13593
Please CHANGE the old X2P logo with the logo with the cross, also please change the information.. I have already paid for this to be done together with the upload with new X2P logo. Please change the old logo and info as people are still buying through old contract. Thanks in advance!