Add files via upload
Please CHANGE the old X2P logo with the logo that is crossed. I have already paid for this to be done together with the upload with new X2P logo. Please change the old logo as people are still buying through old contract and it is confusing. Thanks in advance devs!