Add Hexiar logo and info
Hexiar is a comprehensive service token that combines broker and trading and DeFi. People can trade. We allow users to trade their cryptocurrencies on cryptocurrencies, stocks and even forex. Consumers will also be able to stake their coins, and we are working to introduce this market to the world in the future.