Add Greenheart CBD logo
Continuation of already paid PR trustwallet#14085

• What is the purpose of this token? How will it benefit Trust Wallet users?
$CBD is a utility token that enable customers discounts for our products and also enables our farmer partners to access the funding through our liquidity pools. Our mobile app is able to be integrated with Trust Wallet so Trust Wallet users can use it to get the discount on our products. 
• How is the token being distributed? Are you conducting an airdrop?
The token is a BEP20 token and is already fully distributed. No airdrop.
• Did you sold the tokens at an Initial Coin Offering (ICO) or a token presale?
It was sold as a token presale.
• Is the token listed on a centralized/decentralized exchange?
Yes, it is listed on a decentralized exchange, PancakeSwap.
• Is a Smart Contract audit completed? Where can we find this information?
Yes, an audit is in progress and should be completed soon. https://www.certik.org/projects/greenheart-cbd