BabyDogeZilla Logo Request
Inspired by the coin Dogezilla, we are the first baby of it! We have a strong community and our project has 31,000 holders in just a week. 
Also we are listed in CMC and CG, 2 of the most popular coin site.
We do daily competitions & giveaways. 
Fully community driven token with every day VC with very active people and Mods!
Most of our users uses Trust Wallet, it would be nice if they will see our logo there.
Thank you very much.