Add MTG logo and info on Trust Wallet
Add MTG Logo and Info on Trust Wallet 
MagnetGold (MTG) Token is created by Longreen Corporation Limited. We built MagnetGold (MTG) Token because we strongly believe that it will represent the heart of the ecosystem for a new category of investors, companies seeking to grow their capital, investment funds, venture capital, families, offices, traders and so on. We are on CMC. 
Our Social media links, White Paper, Road Map and other details are on our website: "https://www.mtggold.com/"