Superalgos new asset submission
We are a four-year-old open-source project with tens of thousands of downloads during the beta phase. We need Trust Wallet to make an exception in the 5,000 token holders requirement for a good reason: our token is distributed exclusively among contributors of the open-source project. The current number of holders is counted in the hundreds.

Roadmap: https://superalgos.org/project-roadmap.shtml

Tokenomics: https://superalgos.org/token-overview.shtml