Add BabyDogeZilla Logo To Trust Wallet
We would like to get our project (BabyDogeZilla)'s logo added to Trust Wallet so it displays for investors trying to buy it in the app.  We would also like for the price of our token to display as we are on CMC.  We are willing to pay the submission fee we just need to know where to send the payment as the instructions do not clearly state that part.