Wanting logo added to our CMC listed asset in TW.
Our asset CardanoEvo (cEvo) is currently listed on CoinMarketCap and on Coingecko, yet our logo does not show up in Trustwallet. I reached out to TW support and their email back to me asked me to follow the steps for getting a logo added. So here is our info.json and our logo.png files for addition to the TW database. Thank you in advance. We can be reached at cardanoevo@yahoo.com