Candylad logo and ingo
Candylad was born with the goal of "building a blockchain 4.0 charity community".
Candylad was co-sponsored by a world-renowned charity fund company and the world's top blockchain technology team, and received investment from the richest in Dubai. The company focuses on blockchain technology, smart contracts, ICOs (Candylad) and digital currency investment, focusing on planning to form a charity community, and combining current hotspots such as NFT, DEFI, etc., to provide data for the new charity blockchain to promote public welfare Charity development. At the same time, the project will be launched simultaneously in many countries around the world. The ultimate goal is to create a more open, transparent, reasonable and credible charity platform on a global scale.