Logo and JSON file for TrustWallet
We do not have 10,000 holders just yet as we are a very new project, however, the website says 'This requirement is subject for review based on the project'. We are a real project with doxxed devs, please feel free to review our whitepaper which can be located on our website. 

www.manyworldstoken.com