MONSTAVERSE Token
MonstaVerse is a cryptoverse in itself. An ERC20 token with a set of self-sustaining utilities of its own. The MonstaVerse shall be inclusive of an NFT Marketplace, your one stop shop for everything NFTs. The marketplace would be an NFT listing platform where Ethereum based NFT projects will be provided with a holistic platform to list their NFTs for primary and secondary sales.