Add files via upload
The game will be a PVP shooter game where teams will fight against each other to win. There will be many playable game modes to ensure variety. The players will be able to buy items like weapons and gear with Iron Hunter tokens. There will be organised competitions where players will be able to win rewards such as Iron Hunter tokens. There is also an opportunity for the community to design items for the game and also help with new ideas for the game. We also want to have a NFT market where players will be able to sell their NFT for Iron Hunter tokens.