Add Shooting Star utility BEP20 token.
The purpose of Shooting Star is to create a secured utility BSC-Token [STARS]. Shooting Star’s unique infrastructure allows its believers to invest in a long term project safely. The core team is very experienced. Ways for [STARS] to rocket to the moon are: own launch pad with unique contracts for new projects, unique NFTs and listing Exchanges.[STARS] will be the biggest and brightest BSC Token