Added fura coin. A token backed by emeralds
The  Emerald market is a multi-billion dollar industry.  Thanks to the scarcity and beauty of this precious gemstone.  Whose demand is growing drastically year by year due to its unique green color and rarity. However, the presence of the middleman makes it extremely expensive and difficult to participate in the emerald market, which is the main reason why we created Furacoin. To build a bridge connecting the investors and producers, by eliminating the middleman from the process.