add Doge inu logo and info
Doge Inu is created with the goal of being a decentralized community-driven project with a true purpose. Doge Inu aims to be the most known and most used cryptocurrency in the world. Doge Inu mission is to bring popular cryptocurrency concepts to the mainstream.
Unlike older, comparable projects, Doge Inu introduces holders to next-gen concepts such as participation rewards, NFTs, decentralized exchanges, and more.