sitar token logo for trustwallet
The Sitar token is a BEP-20 token Created on november 11, 2021 on the Binance Smart Chain network. Sitar Token will be a gaming ecosystem to create a new experience in the game world and earn money. In this ecosystem, all family members can play and earn money in a healthy environment with the highest level of security and simplicity.
i want create a logo for this token.