Add RPT logo & Info
Decentralize your financial instruments using the Rupto.
We want to blur the line between the cryptocurrency merchants and ordinary consumers. We will popularize the cryptocurrency by giving free admission, and by helping people understand and confide what we believe to be the future of money. There is a limited supply of coins making it valuable and a great store of wealth like gold. New coins can only be generated by the networks shopping and affiliate programs. So welcome to the future of currency! Safe, fast, secure, anonymous, exciting and most of all fun!