adding a logo and info
CeBioLabs is a young German company with a vision to revolutionize the management of CBD and Cannabis Industry Value Chains. Our team sees itself as a collective of CEOs from the CBD market, Crypto Experts, Economists and Software and System developers. The "CeBioLabs" project, as the issuer of the "CBSL Token”, will develop innovative systems and solutions for the CBD and Cannabis market and launch research projects. All companies wishing to use our systems and solutions must pay the applicable fees in CBSL tokens.