Add Yes World Logo and Info
YesWorld is on the world's biggest mission and it is to save our planet earth. It is mainly working towards the green technology to reduce the carbon footprint. It is working on revolutionary technology to reduce the heat content on the earth’s surface, the one that NASA uses for its astronauts during space programs. YesWorld is striving to bring this revolutionary heat reduction technology in numerous delivery mechanism,  and its perfect amalgamation with the blockchain technology to ensure it reaches to massive scale to mitigate global warming challenges.